+++ app/src/main/java/org/breezyweather/common/source/MainWeatherSource.kt
@@ -50,8 +50,11 @@ interface MainWeatherSource : Source {
      * May be used when you don't have reverse geocoding implemented and you want to filter
      * location results from default location search source to only include some countries
      * for example
+     * @feature When null, it is used as main source
      */
-    fun isWeatherSupportedForLocation(location: Location): Boolean = true
+    fun isFeatureSupportedInMainForLocation(
+        location: Location, feature: SecondaryWeatherSourceFeature? = null
+    ): Boolean = true
 
     /**
      * Returns weather converted to Breezy Weather Weather object