+++ app/src/main/java/org/breezyweather/settings/activities/SettingsActivity.kt
@@ -330,7 +330,7 @@ class SettingsActivity : GeoActivity() {
                         context = this@SettingsActivity,
                         configuredWorldwideSources = sourceManager.getConfiguredMainWeatherSources()
                             .filter {
-                                it.isWeatherSupportedForLocation(Location())
+                                it.isFeatureSupportedInMainForLocation(Location())
                             },
                         configurableSources = sourceManager.getConfigurableSources(),
                         paddingValues = paddings,