+++ app/src/src_nofdroid/org/breezyweather/sources/china/ChinaService.kt
@@ -64,7 +64,10 @@ class ChinaService @Inject constructor(
         SecondaryWeatherSourceFeature.FEATURE_ALERT
     )
 
-    override fun isWeatherSupportedForLocation(location: Location): Boolean {
+    override fun isFeatureSupportedInMainForLocation(
+        location: Location,
+        feature: SecondaryWeatherSourceFeature?
+    ): Boolean {
         return location.countryCode.equals("CN", ignoreCase = true)
     }
 
@@ -121,15 +124,15 @@ class ChinaService @Inject constructor(
     }
 
     // SECONDARY WEATHER SOURCE
-    override val supportedFeatures = listOf(
+    override val supportedFeaturesInSecondary = listOf(
         SecondaryWeatherSourceFeature.FEATURE_AIR_QUALITY,
         SecondaryWeatherSourceFeature.FEATURE_MINUTELY,
         SecondaryWeatherSourceFeature.FEATURE_ALERT
     )
-    override fun isFeatureSupportedForLocation(
-        feature: SecondaryWeatherSourceFeature, location: Location
+    override fun isFeatureSupportedInSecondaryForLocation(
+        location: Location, feature: SecondaryWeatherSourceFeature
     ): Boolean {
-        return isWeatherSupportedForLocation(location)
+        return isFeatureSupportedInMainForLocation(location, feature)
     }
     override val airQualityAttribution = weatherAttribution
     override val pollenAttribution = null