+++ app/src/src_nofdroid/org/breezyweather/sources/eccc/json/EcccAlertSpecialText.kt
@@ -16,10 +16,7 @@
 
 package org.breezyweather.sources.eccc.json
 
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import org.breezyweather.common.serializer.DateSerializer
-import java.util.Date
 
 @Serializable
 data class EcccAlertSpecialText (